First round of optimization
