+++ Dockerfile
@@ -10,4 +10,4 @@ COPY . /app
 
 RUN apt-get update && apt-get install -y libpq-dev gcc python3-psycopg2
 
-RUN pip3 install -r requirements.txt
+RUN asd
