--- config.json.tpl
@@ -1,35 +1,0 @@
-{
-    "command_char": "/",
-    "command_roll": "r",
-    "command_roll_advantage": "v",
-    "command_roll_double_advantage": "vv",
-    "command_roll_disadvantage": "d",
-    "command_dm_roll": "dm",
-    "command_roll_luck": "sorte",
-
-    "command_critical_damage": "cd",
-
-    "command_initiative": "i",
-
-    "command_initiative_adv": "i-v",
-    "command_force_initiative": "i-fixa",
-
-    "command_reset": "i-limpar",
-    "command_remove_initiative": "i-remover",
-
-    "command_add_condition": "i-condicao",
-    "command_remove_condition": "i-condicao-remover",
-
-    "command_next_initiative": "i-p",
-    "command_previous_initiative": "i-a",
-
-    "command_show_stats_player": "my-stats",
-    "command_show_stats_session": "session-stats",
-
-    "command_save_dice": "dice-save",
-    "command_list_saved_dice": "dice",
-    "command_remove_saved_dice": "dice-remover",
-    "command_reset_saved_dice": "dice-limpar",
-
-    "stats": "1"
-}
