+++ docker-compose.yml
@@ -6,6 +6,6 @@ services:
       context: .
       dockerfile: Dockerfile
     volumes:
-      - /var/lib/roll-dices-dnd:/data
+      - /var/lib/roll-dice-dnd:/data
     restart: unless-stopped
-    command: [ "python3", "./discord_client.py" ]
+    command: [ "python3", "./main.py" ]
