--- requirements.txt
@@ -1,7 +1,0 @@
-#discord.py
-discord-py-interactions
-ipython
-python-slugify
-psycopg2-binary==2.9.9
-peewee
-celery
