--- secrets.json.tpl
@@ -1,8 +1,0 @@
-{
-    "discord": "<your_discord_token>",
-    "db": "stats",
-    "user": "postgres",
-    "pass": "",
-    "host": "127.0.0.1",
-    "port": 5432
-}
