+++ src/.env.example
@@ -1,0 +1,9 @@
+discord_token=<your_discord_token>
+db=stats
+user=postgres
+pass=
+host=127.0.0.1
+port=5432
+
+command_char=/
+save_stats_db=1
