+++ src/README.md
@@ -1,0 +1,27 @@
+# Discord bot to roll dice and control initiative table
+
+## Features
+
+
+
+Roll n die 
+
+### Poetry and dependencies
+
+Poetry is a package manager for Python projects. To install it run
+https://python-poetry.org/docs/basic-usage/
+
+```bash
+poetry install
+```
+
+
+### Discord bot
+
+
+
+[comment]: <> (docker compose up -d)
+
+## Running using docker compose
+
+- run the command `docker-compose up -d`
